// src/app/generated-api/model/prestamoRequestDTO.ts
/**
 * API Banco del Cisne
 * API REST para el sistema bancario del Cisne
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@bancodelcisne.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PrestamoRequestDTO { 
    idCliente: number;
    montoSolicitado: number;
    plazoMeses: number;
    tipoPrestamo?: string;
}