// src/app/generated-api/model/pagoPrestamoRequestDTO.ts
/**
 * API Banco del Cisne
 * API REST para el sistema bancario del Cisne
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@bancodelcisne.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PagoPrestamoRequestDTO { 
    idPrestamo: number;
    numeroCuota: number;
    montoPagado: number;
    metodoPago: string;
    referencia?: string;
    mora?: number;
}