// src/app/generated-api/model/amortizacionDTO.ts
/**
 * API Banco del Cisne
 * API REST para el sistema bancario del Cisne
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@bancodelcisne.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AmortizacionDTO { 
    numeroCuota?: number;
    fechaVencimiento?: string;
    montoCuota?: number;
    capital?: number;
    interes?: number;
    saldoRestante?: number;
    estadoPago?: string;
    fechaPago?: string;
}